requirements.txt
Flask
tinydb
requests
python-dotenv
Flask-Cors
streamlit
pandas
pydeck